requirements.txt
fastapi
gunicorn
uvicorn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx